Update Fan_Feng_Project_One
The idea of my generative visual system is to try the possibility of cooperate with music students. The object of a keyboard in Max MSP is calling Kslider, the object name shows it is also a slider, and it will work like a slider. It means people can use the keyboard on to change the value of different controllable parameters when the Kslider connected with a midi controller. When the music students use the system, they will able to change the visual by play music with play keyboard, and the visual will change with their rhythm and beat. So, the music student can more focus on playing music. In this system, I am letting people use the keyboard to change the color tone of the video. I connect the kslider to a slider, then connect the slider to a vizzie effect object that can change the color tone. There are some other effects will be mixing with the color tone changing effect. I fine out the number of the dials on my midi controller, and connect this dials to the dials objects in Max MSP. People can use it to change the value of the effects and video mixers.