Description of Project Two
In the this project two, I had to used limited object in MAX MSP to create an audio visual artwork. In the beginning, I feel the object I can use is very few, I even cannot use most of the drawing relate object to generate image of visual content. So, I tried to create a work as less kinds of object as possible. I use numbers of sliders to create a “mulit” slider, and let them generate the wave randomly. I create a drunk object to generate random number for the first sliders, then I used pipe and delay object to transport the number to the next slider with a few delays, after I did this action many time, the number random generated will transport from the beginning to the end slider by slider and create a wave. The number from the first slider will use for make note for the audio part of this work.  This work doesn’t have to doing much operating for running it, there are only on trigger and on slider in the presentation mode. The choose orange and white for the main tone because they have high contrast but still look comfortable.