Create Fan_Feng_readning_three
I feed I am reading a fake class reading when I was reading it, this reading is a handbook for the audience, but not an academic article. This handbook is an opera of New York philharmonic, most parts of this handbook are the dialogue of the opera and the introduction of some of the philharmonic members. The opera is a story from the old testament call Elijah, the dialogue seems mostly like the bible. There is nothing special in the story. The dialogue part of this handbook calls Texts and Translations, it shows the opera maybe sing in another language than English. This is a very interesting idea for a music concert. People usually said there is not a border in music, but the audience may feel boring if they don’t understand the dialogue. Translations on the handbook will help the audience to understand the performance. In the other hand, more and more music artist play with the digital music. Some digital effect may make singer’s voice different to listen, the text will help the audience to know what the singer sings. The text also tells the audience what kind of music it is playing, such as the Trio, Chorus, Aria etc. Digital art is still unfamiliar to most of the people. I in a music performance with Digital art, some desorption text in the handbook may help the audience to understand what is happening om the performance. As a digital artist, we may use some media other than a paper handbook to guiding the audience, it maybe meets the theme of digital art better. There is a small instrumentation part in the handbook about the instrument used in this opera. In seems not a very important part of the opera or music concert because the audience will see the instrument in the performance. In digital art performance, there are many things hiding in the computer, there are different visual or audio software use in the performance, and there are also different setup and programming in one software. The audience is very curious, they always want to know how performer doing something on the stage in the performance. Written a few technique information in the handbook will meet the audience’s curiosity, and it will also help them understand the performance. It doesn’t have to a very detail introduction, a few basic information will help already.